Refactor, reorganization, new tests
