+++ request-manager/spec/errors.go
@@ -4,7 +4,16 @@ package spec
 
 import (
 	"fmt"
+	"strings"
 )
+
+func stringSetToArray(set map[string]bool) []string {
+	arr := []string{}
+	for v, _ := range set {
+		arr = append(arr, v)
+	}
+	return arr
+}
 
 /* =========================================================================== */
 
@@ -14,7 +23,7 @@ type InvalidValueError struct {
 	Sequence string
 	Node     *string
 	Field    string
-	Value    string
+	Values   []string
 	Expected string
 }
 
@@ -27,8 +36,11 @@ func (e InvalidValueError) Error() string {
 		loc = fmt.Sprintf("sequence %s, node %s", e.Sequence, *e.Node)
 	}
 
-	return fmt.Sprintf("%s, field `%s` contains invalid value \"%s\", expected %s",
-		loc, e.Field, e.Value, e.Expected)
+	var values string
+	values = fmt.Sprintf("\"%s\"", strings.Join(e.Values, "\", \""))
+
+	return fmt.Sprintf("%s: invalid value(s) %s in field `%s`, expected %s",
+		loc, values, e.Field, e.Expected)
 }
 
 /* =========================================================================== */
@@ -59,7 +71,7 @@ func (e MissingValueError) Error() string {
 		explanation = fmt.Sprintf(": %s", e.Explanation)
 	}
 
-	return fmt.Sprintf("%s, field `%s` missing%s", loc, e.Field, explanation)
+	return fmt.Sprintf("%s: field(s) `%s` missing%s", loc, e.Field, explanation)
 }
 
 /* =========================================================================== */
@@ -67,10 +79,11 @@ func (e MissingValueError) Error() string {
 var _ error = DuplicateValueError{}
 
 type DuplicateValueError struct {
-	Sequence string
-	Node     *string
-	Field    string
-	Value    string
+	Sequence    string
+	Node        *string
+	Field       string
+	Values      []string
+	Explanation string
 }
 
 func (e DuplicateValueError) Error() string {
@@ -82,6 +95,17 @@ func (e DuplicateValueError) Error() string {
 		loc = fmt.Sprintf("sequence %s, node %s", e.Sequence, *e.Node)
 	}
 
-	return fmt.Sprintf("%s, field `%s` contains duplicate value \"%s\"",
-		loc, e.Field, e.Value)
+	var values string
+	values = fmt.Sprintf("\"%s\"", strings.Join(e.Values, "\", \""))
+
+	var explanation string
+	switch e.Explanation {
+	case "":
+		explanation = ""
+	default:
+		explanation = fmt.Sprintf(": %s", e.Explanation)
+	}
+
+	return fmt.Sprintf("%s: value(s) %s duplicated in field `%s`%s",
+		loc, values, e.Field, explanation)
 }
