+++ request-manager/test/specs/fail-args-each-not-renamed-node-check.yaml
@@ -4,10 +4,13 @@ sequences:
     args:
       required:
         - name: required-a
-    acl:
-        - ops: ["start"]
     nodes:
       node-a:
         category: job
         type: job-type-a
+        each:
+          - arg-a:alias-a
+        args:
+          - expected: arg-x
+            given: arg-a # arg-a renamed as alias-a and as arg-x
         deps: []
