+++ request-manager/test/specs/fail-args-each-once-node-check.yaml
@@ -4,10 +4,13 @@ sequences:
     args:
       required:
         - name: required-a
-    acl:
-        - ops: ["start"]
     nodes:
       node-a:
         category: job
         type: job-type-a
+        each:
+          - arg-a:alias-a
+        args:
+          - expected: alias-a # can't have arg-a and arg-b be renamed as alias-a
+            given: arg-b
         deps: []
