+++ request-manager/test/specs/fail-no-duplicate-args-sequence-check.yaml
@@ -3,9 +3,10 @@ sequences:
   seq-a:
     args:
       required:
-        - name: required-a
-    acl:
-        - ops: ["start"]
+        - name: arg-a
+      optional:
+        - name: arg-a # sequence args can't be duplicated
+          default: ""
     nodes:
       node-a:
         category: job
